🔨 refactor: #84 Move account settings into new class
