+++ src/settings/index.ts
@@ -1,0 +1,3 @@
+export type { GithubAccount, GithubLinkPluginData, GithubLinkPluginSettings } from "./types";
+export { DEFAULT_SETTINGS } from "./types";
+export { GithubLinkPluginSettingsTab } from "./settings-tab";
